pyproject.toml
[tool.poetry]
name = "flask-template-app"
version = "0.1.0"
description = ""
authors = ["Alexander Gerbik <alexander.gerbik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.0"
blinker = "^1.4"
flask = "^1.1.1"
gunicorn = "^20.0.4"
marshmallow = "^3.4.0"
psycopg2-binary = "^2.8.4"
pyjwt = {extras = ["crypto"], version = "^1.7.1"}
sqlalchemy = "^1.3.13"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
freezegun = "^0.3.14"
isort = "^4.3.21"
logging-tree = "^1.8.1"
migra = "^1.0.1575954682"
pre-commit = "^2.0.1"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
pytest-mock = "^2.0.0"

[tool.isort]
line_length=88
multi_line_output=3
include_trailing_comma=true
combine_as_imports=true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
